Fixed errors during testing

+++ services/madoc-ts/src/utility/sync-jwt-requests.ts
@@ -1,4 +1,4 @@
-import { readdirSync, readFileSync, writeFileSync } from 'fs';
+import { readdirSync, readFileSync, existsSync, writeFileSync } from 'fs';
 import * as path from 'path';
 import { generateServiceToken } from './generate-service-token';
 
@@ -8,11 +8,15 @@ export async function syncJwtRequests() {
   const directory = readdirSync(jwtRequests);
   for (const file of directory) {
     if (file.endsWith('.json')) {
-      const request = readFileSync(path.join(jwtRequests, file)).toString('utf-8');
-      const json = JSON.parse(request);
-      const token = await generateServiceToken(json);
-      if (token) {
-        writeFileSync(path.join(jwtResponses, file), `{"token": "${token}"}`);
+      const dest = path.join(jwtResponses, file);
+
+      if (!existsSync(dest)) {
+        const request = readFileSync(path.join(jwtRequests, file)).toString('utf-8');
+        const json = JSON.parse(request);
+        const token = await generateServiceToken(json);
+        if (token) {
+          writeFileSync(path.join(jwtResponses, file), `{"token": "${token}"}`);
+        }
       }
     }
   }
